+++ CHANGELOG.md
@@ -4,20 +4,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-## [0.5.1.0]
+## [Unreleased]## [0.5.2.0b]
 ### Changed
-- data asset folder location in CP4D v4.0.2 (with new git)
-- save Excel/lp/csv files using ibm_watson_studio_lib in CPD v4.0.2 in ScenarioManager
+- (minor) Release dates in this change log
+- ScenarioDbManager support for DB2 in cloud
+- Fixed AutoScenarioDbTable functionality in ScenarioDbManager
+- Bumped-up version to 0.5.2 due to many changes
+### Added
+- Optional forced platform/version in ScenarioManager
+- DataManager.print_inputs_outputs_summary() method
+- inputs and outputs arguments to ScenarioManager.__init__
+- Mixed-case db_table_name warning in ScenarioDbTable
+- DB-table-name reserved word warning in ScenarioDbTable
+- plotlymanager module
+- Cached read of scenarios table in ScenarioDbManager
+### Removed
+- (minor) Removed support for DSX as platform choice
+
+## [0.5.1.0] - 2021-11-30
+### Changed
+- Writing data asset in CP4DaaS (as of 30 Nov 2021)
+- Data asset folder location in CP4D v4.0.2 (with new git)
+- Save Excel/lp/csv files using ibm_watson_studio_lib in CPD v4.0.2 in ScenarioManager
 ### Added
 - DataManager.extract_solution static method
 - DataManager.get_raw_table_by_name method
 
-## [0.5.0.1]
+## [0.5.0.1] - 2021-10-29
 ### Changed
 - ScenarioManager.write_data_into_scenario now supports a template scenario name when creating a new scenario.
 
-## [0.5.0.0]
+## [0.5.0.0] - 2021-08-14
 ### Changed
 - Replace dd-scenario with decision-optimization-client==1.0.0 package (essential for CP4D 4.0)
 - Deprecated the DOModelExporter: dev does not support exporting a DO model in any way.
@@ -25,7 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - 
 
-## [0.4.1.0]
+## [0.4.1.0] - 2021-07-16
 ### Added
 - MultiScenarioManager
 - ScenarioManager.load_data_from_excel_s load subset of tables